The Bead sheet export function now exports higher quality pngs.
The bead sheet quality was very low. When printing, it was almost
impossible to read the color names. This commit adds a simple
gen-chars script that generates the required alphanum characters
with some very simple svg code. Inkscape is used to convert the
svg to png. A set of higher quality alphanum pngs has already
been generated, and the necessary variables have been changed in
the file AbstractReduceColorsFilter.java in order to produce
much higher quality bead sheets.

Signed-off-by: Vangelis Tasoulas <vangelis@tasoulas.net>