Create Hama-mini-builtin
All new colors for Hama mini beads.
I just found out that as far as I can tell, the above are the same currently for the midi beads.
So, you'd be better off overwriting the hama-builtin with this list.
I used Photoshop to get the colors from the website, but I have no clue if this will be the best way to get the right colors.
They seem to look right. Haven't changed the ones that were already added  by you.
Flesh has changed from 25 to 26 btw.